feat(evaluator): consume messages from grouper queue
VULN-2627
