+++ evaluator/common.py
@@ -52,9 +52,9 @@ class EvaluatorException(Exception):
     """Evaluator processing exception"""
 
 
-class EvaluatorMessageTypes(Enum):
+class EvaluatorMessageType(Enum):
     """Message types which can arrive at kafka"""
-    UPLOAD_NEW_FILE = "upload_new_file"
+    EVALUATE_SYSTEM = "upload_new_file"
     RE_EVALUATE_SYSTEM = "re-evaluate_system"
