+++ evaluator/evaluator.py
@@ -28,16 +28,16 @@ DB_POOL = None
 
 # prometheus probes
 # times
-VMAAS_EVAL_TIME = Histogram('ve_evaluator_vmaas_evaluation_seconds', 'Time spent checking a system for vmaas hits',
+VMAAS_EVAL_TIME = Histogram('ve_eval_vmaas_eval_seconds', 'Time spent checking a system for vmaas hits',
                             buckets=[0.1, 0.25, 0.5, 0.75, 1.0, 1.5, 2, 5])
 # counts
-VMAAS_COUNT = Counter('ve_evaluator_vmaas_calls', 'Number of VMaaS-evaluations attempted')
-INV_ID_NOT_FOUND = Counter('ve_evaluator_inventory_not_found', 'Number of times inventory-id not in SystemPlatform')
-UNKNOWN_MSG = Counter('ve_evaluator_unknown_msg', 'Number of unrecognized messages delivered from queue')
-UNKNOWN_TOPIC = Counter('ve_evaluator_unknown_topic', 'Number of times message delivered from unsupported topic')
-MESSAGE_PARSE_ERROR = Counter('ve_evaluator_message_parse_error', '# of message parse errors')
-VMAAS_ERRORS_SKIP = Counter('ve_evaluator_vmaas_errors_skip', '# of evaluations skipped due to VMaaS errors')
-STORE_FAILED = Counter('ve_evaluator_store_failed', '# of failed evaluations')
+VMAAS_COUNT = Counter('ve_eval_vmaas_calls', 'Number of VMaaS-evaluations attempted')
+INV_ID_NOT_FOUND = Counter('ve_eval_inventory_not_found', 'Number of times inventory-id not in SystemPlatform')
+UNKNOWN_MSG = Counter('ve_eval_unknown_msg', 'Number of unrecognized messages delivered from queue')
+UNKNOWN_TOPIC = Counter('ve_eval_unknown_topic', 'Number of times message delivered from unsupported topic')
+MESSAGE_PARSE_ERROR = Counter('ve_eval_message_parse_error', '# of message parse errors')
+VMAAS_ERRORS_SKIP = Counter('ve_eval_vmaas_errors_skip', '# of evaluations skipped due to VMaaS errors')
+STORE_FAILED = Counter('ve_eval_store_failed', '# of failed evaluations')
 
 CONSUMER_QUEUE = mqueue.MQReader(CFG.evaluator_topics)
 PAYLOAD_TRACKER_PRODUCER = mqueue.MQWriter(CFG.payload_tracker_topic)
